pyproject.toml
[tool.poetry]
name = "MK19-EarthInvader"
version = "1.3"
description = "Discord bot using the cryptography toolkit of SkeletonKey"
authors = ["Soulsender", "0xSolanaceae", "FyzzLive"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
Cryptography = "^43.0.1"
python-dotenv = "~1.0.1"
nextcord = "^2.6.0"
aiohttp = "^3.10.9"
discord_webhook = "^1.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"